+++ packages/core/src/agent/certificate.test.ts
@@ -13,8 +13,10 @@ import { type Agent, HttpAgent, IC_ROOT_KEY } from './agent/index.ts';
 import {
   CertificateHasTooManyDelegationsErrorCode,
   CertificateNotAuthorizedErrorCode,
+  CertificateNotAuthorizedForSubnetErrorCode,
   CertificateTimeErrorCode,
   CertificateVerificationErrorCode,
+  LookupErrorCode,
   ProtocolError,
   TrustError,
   UnexpectedErrorCode,
@@ -23,6 +25,7 @@ import {
 } from './errors.ts';
 import { utf8ToBytes, hexToBytes, bytesToHex } from '@noble/hashes/utils';
 import { uint8Equals } from './utils/buffer.ts';
+import { goldenCertificates } from './agent/http/__certificates__/goldenCertificates.ts';
 
 const MINUTES_TO_MSEC = 60 * 1000;
 
@@ -596,7 +599,7 @@ test('delegation works for canisters within the subnet range', async () => {
       Cert.Certificate.create({
         certificate: SAMPLE_CERT_BYTES,
         rootKey: hexToBytes(IC_ROOT_KEY),
-        canisterId,
+        principal: { canisterId },
         blsVerify: async () => true,
       }),
     ).resolves.not.toThrow();
@@ -620,7 +623,7 @@ test('delegation check fails for canisters outside of the subnet range', async (
       await Cert.Certificate.create({
         certificate: SAMPLE_CERT_BYTES,
         rootKey: hexToBytes(IC_ROOT_KEY),
-        canisterId: canisterId,
+        principal: { canisterId },
       });
     } catch (error) {
       expect(error).toBeInstanceOf(TrustError);
@@ -644,7 +647,7 @@ test('certificate verification fails for an invalid signature', async () => {
     await Cert.Certificate.create({
       certificate: badCertEncoded,
       rootKey: hexToBytes(IC_ROOT_KEY),
-      canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai'),
+      principal: { canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai') },
     });
   } catch (error) {
     expect(error).toBeInstanceOf(TrustError);
@@ -660,7 +663,7 @@ test('certificate verification fails if the time of the certificate is > 5 minut
     await Cert.Certificate.create({
       certificate: SAMPLE_CERT_BYTES,
       rootKey: hexToBytes(IC_ROOT_KEY),
-      canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai'),
+      principal: { canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai') },
       blsVerify: async () => true,
     });
   } catch (error) {
@@ -683,7 +686,7 @@ test('certificate creation passes if the time of the certificate is > 5 minutes 
   const cert = await Cert.Certificate.create({
     certificate: SAMPLE_CERT_BYTES,
     rootKey: hexToBytes(IC_ROOT_KEY),
-    canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai'),
+    principal: { canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai') },
     blsVerify: async () => true,
     disableTimeVerification: true,
   });
@@ -701,7 +704,7 @@ test('certificate creation fails if the time of the certificate is > 5 minutes i
     await Cert.Certificate.create({
       certificate: SAMPLE_CERT_BYTES,
       rootKey: hexToBytes(IC_ROOT_KEY),
-      canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai'),
+      principal: { canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai') },
       blsVerify: async () => true,
       agent: {} as Agent,
     });
@@ -734,7 +737,7 @@ test('certificate creation passes if the time of the certificate is > 5 minutes 
   const cert = await Cert.Certificate.create({
     certificate: SAMPLE_CERT_BYTES,
     rootKey: hexToBytes(IC_ROOT_KEY),
-    canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai'),
+    principal: { canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai') },
     blsVerify: async () => true,
     agent,
   });
@@ -762,7 +765,7 @@ test('certificate creation passes if the time of the certificate is > 5 minutes 
   const cert = await Cert.Certificate.create({
     certificate: SAMPLE_CERT_BYTES,
     rootKey: hexToBytes(IC_ROOT_KEY),
-    canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai'),
+    principal: { canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai') },
     blsVerify: async () => true,
     agent,
   });
@@ -789,7 +792,7 @@ test('certificate creation fails if the time of the certificate is > 5 minutes i
     await Cert.Certificate.create({
       certificate: SAMPLE_CERT_BYTES,
       rootKey: hexToBytes(IC_ROOT_KEY),
-      canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai'),
+      principal: { canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai') },
       blsVerify: async () => true,
       agent,
     });
@@ -831,7 +834,7 @@ test('certificate creation fails if the time of the certificate is > max age min
     await Cert.Certificate.create({
       certificate: SAMPLE_CERT_BYTES,
       rootKey: hexToBytes(IC_ROOT_KEY),
-      canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai'),
+      principal: { canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai') },
       blsVerify: async () => true,
       maxAgeInMinutes,
       agent,
@@ -853,7 +856,7 @@ test('certificate verification fails if the time of the certificate is > 5 minut
     await Cert.Certificate.create({
       certificate: SAMPLE_CERT_BYTES,
       rootKey: hexToBytes(IC_ROOT_KEY),
-      canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai'),
+      principal: { canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai') },
       blsVerify: async () => true,
     });
   } catch (error) {
@@ -869,7 +872,7 @@ test('certificate creation passes if the time of the certificate is > 5 minutes 
   const cert = await Cert.Certificate.create({
     certificate: SAMPLE_CERT_BYTES,
     rootKey: hexToBytes(IC_ROOT_KEY),
-    canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai'),
+    principal: { canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai') },
     blsVerify: async () => true,
     disableTimeVerification: true,
   });
@@ -884,7 +887,7 @@ test('certificate verification fails if the time of the certificate is > 5 minut
     await Cert.Certificate.create({
       certificate: SAMPLE_CERT_BYTES,
       rootKey: hexToBytes(IC_ROOT_KEY),
-      canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai'),
+      principal: { canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai') },
       blsVerify: async () => true,
       agent: {} as Agent,
     });
@@ -917,7 +920,7 @@ test('certificate creation passes if the time of the certificate is > 5 minutes 
   const cert = await Cert.Certificate.create({
     certificate: SAMPLE_CERT_BYTES,
     rootKey: hexToBytes(IC_ROOT_KEY),
-    canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai'),
+    principal: { canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai') },
     blsVerify: async () => true,
     agent,
   });
@@ -945,7 +948,7 @@ test('certificate creation passes if the time of the certificate is > 5 minutes 
   const cert = await Cert.Certificate.create({
     certificate: SAMPLE_CERT_BYTES,
     rootKey: hexToBytes(IC_ROOT_KEY),
-    canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai'),
+    principal: { canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai') },
     blsVerify: async () => true,
     agent,
   });
@@ -975,7 +978,7 @@ test('certificate creation fails if the time of the certificate is > 5 minutes i
     await Cert.Certificate.create({
       certificate: SAMPLE_CERT_BYTES,
       rootKey: hexToBytes(IC_ROOT_KEY),
-      canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai'),
+      principal: { canisterId: Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai') },
       blsVerify: async () => true,
       agent,
     });
@@ -1014,7 +1017,7 @@ test('certificate verification fails on nested delegations', async () => {
     await Cert.Certificate.create({
       certificate: overlyNested,
       rootKey: hexToBytes(IC_ROOT_KEY),
-      canisterId: canisterId,
+      principal: { canisterId },
     });
   } catch (error) {
     expect(error).toBeInstanceOf(ProtocolError);
@@ -1026,7 +1029,7 @@ test('certificate verification fails on nested delegations', async () => {
     await Cert.Certificate.create({
       certificate: overlyNested,
       rootKey: hexToBytes(IC_ROOT_KEY),
-      canisterId: canisterId,
+      principal: { canisterId },
     });
   } catch (error) {
     expect(error).toBeInstanceOf(ProtocolError);
@@ -1034,4 +1037,117 @@ test('certificate verification fails on nested delegations', async () => {
       CertificateHasTooManyDelegationsErrorCode,
     );
   }
+});
+
+test('certificate verification fails if a subnet ID is provided as the principal', async () => {
+  jest.setSystemTime(parseTimeFromCert(SAMPLE_CERT_BYTES));
+
+  expect.assertions(2);
+  try {
+    await Cert.Certificate.create({
+      certificate: SAMPLE_CERT_BYTES,
+      rootKey: hexToBytes(IC_ROOT_KEY),
+      principal: {
+        subnetId: Principal.fromText(
+          'utdle-wwxpm-vc64m-zxguk-5sj74-35vrb-tbgwg-pcird-5gr2x-52oxl-bea',
+        ),
+      },
+    });
+  } catch (error) {
+    expect(error).toBeInstanceOf(TrustError);
+    expect((error as TrustError).cause.code).toBeInstanceOf(
+      CertificateNotAuthorizedForSubnetErrorCode,
+    );
+  }
+});
+
+describe('subnet certificates', () => {
+  const certBytes = hexToBytes(goldenCertificates.mainnetApplicationO3ow2);
+  const certTime = parseTimeFromCert(certBytes).getTime();
+  const subnetId = Principal.fromText(
+    'o3ow2-2ipam-6fcjo-3j5vt-fzbge-2g7my-5fz2m-p4o2t-dwlc4-gt2q7-5ae',
+  );
+
+  beforeEach(() => {
+    jest.setSystemTime(certTime);
+  });
+
+  it('certificate creation passes if the certificate is from a subnet', async () => {
+    const cert = await Cert.Certificate.create({
+      certificate: certBytes,
+      rootKey: hexToBytes(IC_ROOT_KEY),
+      principal: { subnetId },
+    });
+    expect(cert).toBeInstanceOf(Cert.Certificate);
+  });
+
+  it('certificate creation fails if the certificate is from another subnet', async () => {
+    const anotherSubnetId = Principal.fromText(
+      'uzr34-akd3s-xrdag-3ql62-ocgoh-ld2ao-tamcv-54e7j-krwgb-2gm4z-oqe',
+    );
+
+    expect.assertions(2);
+    try {
+      await Cert.Certificate.create({
+        certificate: certBytes,
+        rootKey: hexToBytes(IC_ROOT_KEY),
+        principal: { subnetId: anotherSubnetId },
+      });
+    } catch (error) {
+      expect(error).toBeInstanceOf(TrustError);
+      expect((error as TrustError).cause.code).toBeInstanceOf(
+        CertificateNotAuthorizedForSubnetErrorCode,
+      );
+    }
+  });
+
+  test('certificate verification fails if the time of the certificate is > 5 minutes in the past', async () => {
+    const tenMinutesFuture = parseTimeFromCert(SAMPLE_CERT_BYTES).getTime() + 10 * MINUTES_TO_MSEC;
+    jest.setSystemTime(new Date(tenMinutesFuture));
+
+    expect.assertions(2);
+    try {
+      await Cert.Certificate.create({
+        certificate: certBytes,
+        rootKey: hexToBytes(IC_ROOT_KEY),
+        principal: { subnetId },
+      });
+    } catch (error) {
+      expect(error).toBeInstanceOf(TrustError);
+      expect((error as TrustError).cause.code).toBeInstanceOf(CertificateTimeErrorCode);
+    }
+  });
+
+  it('certificate verification fails if the time of the certificate is > 5 minutes in the future', async () => {
+    const tenMinutesPast = certTime - 10 * MINUTES_TO_MSEC;
+    jest.setSystemTime(new Date(tenMinutesPast));
+
+    expect.assertions(2);
+    try {
+      await Cert.Certificate.create({
+        certificate: certBytes,
+        rootKey: hexToBytes(IC_ROOT_KEY),
+        principal: { subnetId },
+      });
+    } catch (error) {
+      expect(error).toBeInstanceOf(TrustError);
+      expect((error as TrustError).cause.code).toBeInstanceOf(CertificateTimeErrorCode);
+    }
+  });
+
+  it('certificate verification fails if a canister ID is provided as the principal', async () => {
+    const canisterId = Principal.fromText('ivg37-qiaaa-aaaab-aaaga-cai');
+
+    expect.assertions(2);
+    try {
+      await Cert.Certificate.create({
+        certificate: certBytes,
+        rootKey: hexToBytes(IC_ROOT_KEY),
+        principal: { canisterId },
+      });
+    } catch (error) {
+      expect(error).toBeInstanceOf(ProtocolError);
+      expect((error as ProtocolError).cause.code).toBeInstanceOf(LookupErrorCode);
+    }
+  });
 });