+++ CHANGELOG.md
@@ -8,10 +8,16 @@
   - Renames `isV3ResponseBody` to `isV4ResponseBody`
   - Renames `HttpV3ApiNotSupportedErrorCode` to `HttpV4ApiNotSupportedErrorCode`
 - feat(agent)!: use `/api/v3` for query and read_state requests
+- feat(agent)!: supports both subnet id and canister id for certificate verification
+  - The `canisterId` option has been replaced with the `principal` option in the `Certificate.create` options object
 - feat(assets)!: replaces `@dfinity/{agent,candid,principal}` deps with `@icp-sdk/core`
 - feat(assets)!: drops support for cjs for the `@dfinity/assets` package
 - feat(auth-client)!: `@dfinity/auth-client` has been deprecated. Migrate to [`@icp-sdk/auth`](https://js.icp.build/auth/latest/upgrading/v4)
 - feat(agent): lookup canister ranges using the `/canister_ranges/<subnet_id>/<ranges>` certificate path
+- feat(agent): introduce the `lookupCanisterRanges`, `lookupCanisterRangesFallback`, and `decodeCanisterRanges` utility functions to lookup canister ranges from certificate trees
+- feat(agent): introduce the `getSubnetIdFromCanister` and `readSubnetState` methods in the `HttpAgent` class
+- feat(agent): introduce the `SubnetStatus` utility namespace to request subnet information directly from the IC public API
+- feat(agent): export `IC_STATE_ROOT_DOMAIN_SEPARATOR` constant
 - refactor(agent): only declare IC URLs once in the `HttpAgent` class
 - refactor(agent): split inner logic of `check_canister_ranges` into functions
 - test(principal): remove unneeded dependency